댓글 구현 완료2

+++ src/main/java/com/leeacademy/board/controller/BoardController.java
@@ -5,11 +5,12 @@ import com.leeacademy.board.repository.BoardRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.support.RequestContextUtils;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.Map;
 
 @Controller
 public class BoardController {
@@ -49,7 +50,11 @@ public class BoardController {
 	}
 
 	@GetMapping("/edit")
-	public String edit(@RequestParam long no, Model model) {
+	public String edit(HttpServletRequest request, Model model) {
+		Map<String, ?> map = RequestContextUtils.getInputFlashMap(request);
+		if(map == null) throw new RuntimeException("권한 없음");
+		Long no = (Long)map.get("no");
+
 		Board board = boardRepository.findById(no).orElseThrow();
 		model.addAttribute("board", board);
 		return "edit";
@@ -61,14 +66,41 @@ public class BoardController {
 		origin.setTitle(board.getTitle());
 		origin.setWriter(board.getWriter());
 		origin.setContent(board.getContent());
+//		비밀번호 추가
+		origin.setPassword(board.getPassword());
 		Board result = boardRepository.save(origin);
 		attr.addAttribute("no", result.getNo());
 		return "redirect:detail";
 	}
 
 	@GetMapping("/delete")
-	public String delete(@RequestParam long no){
+	public String delete(HttpServletRequest request){
+
+		Map<String, ?> map = RequestContextUtils.getInputFlashMap(request);
+		if(map == null) throw new RuntimeException("권한 없음");
+		Long no = (Long)map.get("no");
+
 		boardRepository.deleteById(no);
 		return "redirect:/";
 	}
+
+	@GetMapping("/password/{mode:edit|delete}/{no}")
+	public String password(@PathVariable String mode, @PathVariable int no, Model model) {
+		model.addAttribute("mode",mode);
+		model.addAttribute("no",no);
+		return "password";
+	}
+
+	@PostMapping("/password/{mode:edit|delete}/{no}")
+	public String password(@PathVariable String mode, @PathVariable long no, @RequestParam String password,
+						   RedirectAttributes attr) {
+		Board board = boardRepository.findById(no).orElseThrow();
+		if(board.getPassword().equals(password)) {
+			attr.addFlashAttribute("no",no);
+			return "redirect:/" + mode;
+		} else {
+			return "redirect:/password/" + mode + "/" + no + "?error";
+		}
+	}
+
 }
